renamed vertexActions to graphActions
